Update chicago_bus_route_info.py
This uses Chicago public transportation website, and gives a live representation of when a specific bus on a specified route, arrives at each stop. The web page is analyzed as an XML document in which I extract the information I need, and write the data to an XML document. 
Utilizing urllib.request module and file handling.